Don't close audio streams until we're sure we've found a file, and always attempt to load the requested file if the stream is closed.
This matches RGSS's behavior should an Audio method be called for a non-existent file, and prevents a segfault if an attempt raises an error and then play is called again for the previously playing file.

As part of this fix, we're now storing the SDL_RWops for audio streams directly in the source structs.