Avoid out-of-bounds array access in tilemap shader
This breaks the shader on my Android device. When `atIndex` is out of
bounds, `atFrames[atIndex]` evaluates to 0 and so
`aniIndex / atFrames[atIndex]` results in division by 0. All of the
tiles where this division by 0 occurs are blank on my Android device.
That probably means the shader pipeline halts when division by 0 occurs.