Throttle the frame rate in libretro builds with threaded audio
When not using the threaded audio driver, the frontend is supposed to
block until enough audio samples have been rendered for one frame, so
the frontend throttles the frame rate automatically if the core's frame
rate is slower than the display frame rate.

However, when using the threaded audio driver, there's no more
throttling, so we have to do it ourselves.