Change .mkxp file extension to not be recognized as an archive in libretro builds
I'm planning on allowing the user to put an empty file with the
extension .mkxp in the game directory and load that as the game so that
they can control the file name RetroArch uses for save states, since
RetroArch uses the name of the game file as the name of the save states.

The file extension .mkxpz is still recognized as an archive.