Remove base ruby build from libretro/Makefile
Stage 1 of the libretro build process originally built a base Ruby
targeting the build machine that was then used to build the version of
Ruby actually used in the core for better determinism, but it turns out
that the base Ruby matters very little. We can just use the system
version of Ruby as the base Ruby.